# Conflicts:
#	src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineRepositoryService.kt
#	src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/PipelineVersionFacadeService.kt
#	src/backend/ci/core/store/api-store/src/main/kotlin/com/tencent/devops/store/api/atom/ServiceMarketAtomEnvResource.kt

+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/common/utils/AgentSystemInfoUtils.kt
@@ -117,7 +117,7 @@ object AgentSystemInfoUtils {
                 name = network.name
                 displayName = network.displayName
                 macAddr = network.macaddr
-                mtu = network.mtu
+                mtu = network.mtu.toInt()
                 ipv4Addr = network.iPv4addr.toList()
                 ipv6Addr = network.iPv6addr.toList()
                 speed = network.speed
